refactor: update imports after feature-based restructuring
