+++ frontend/src/components/layout/ProtectedRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '../../hooks/useAuth';
+import { useAuth } from '../../features/auth/hooks/useAuth';
 
 const ProtectedRoute = ({ children }) => {
   const { isAuthenticated } = useAuth();
